libc: Move syslog_stream from nuttx/streams.h to drivers/syslog/syslog.h
